refactor : 소스 코드 정리
