+++ src/main/java/racingcar/domain/validator/NumberValidator.java
@@ -22,8 +22,4 @@ public class NumberValidator {
 
         return result;
     }
-
-    public int validate(int number) {
-        return validate(number + "");
-    }
 }
